+++ fancy_dataclass/__init__.py
@@ -10,7 +10,7 @@ from .subprocess import SubprocessDataclass
 from .toml import TOMLDataclass
 
 
-__version__ = '0.6.0'
+__version__ = '0.6.1'
 
 __all__ = [
     'ArgparseDataclass',